Add Journal management functionality with CRUD operations and error handling improvements

+++ backend/src/modules/Dev.Module.Accounting/Domain/Entities/JournalEntry.cs
@@ -8,8 +8,8 @@ public class JournalEntry
     public DateTime EntryDate { get; set; }
     public string Reference { get; set; } = string.Empty;
     public string Description { get; set; } = string.Empty;
-    public float TotalDebit { get; set; } = 0;
-    public float TotalCredit { get; set; } = 0;
+    public decimal TotalDebit { get; set; } = 0;
+    public decimal TotalCredit { get; set; } = 0;
     /// <summary>
     /// 0 draft, 1 posted, 2 cancelled
     /// </summary>
