+++ backend/src/modules/Dev.Module.Accounting/Domain/Entities/JournalEntryLine.cs
@@ -6,7 +6,7 @@ public class JournalEntryLine
     public Guid JournalEntryId { get; set; }
     public JournalEntry JournalEntry { get; set; } = default!;    
     public Guid AccountId { get; set; }
-    public float Debit { get; set; } = 0;
-    public float Credit { get; set; } = 0;
+    public decimal Debit { get; set; } = 0;
+    public decimal Credit { get; set; } = 0;
     public string Description { get; set; } = string.Empty;
 }
